day08/ex03/requirements.txt
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.6
google-api-core==2.1.0
google-auth==2.3.0
google-cloud-bigquery==2.28.1
google-cloud-core==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpcio==1.41.0
idna==3.2
packaging==21.0
pkg_resources==0.0.0
proto-plus==1.19.3
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
requests==2.26.0
rsa==4.7.2
six==1.16.0
urllib3==1.26.7